Added .env with base API prefix for easier reference
New store to manage student operations. Added fetchStudent action to use API endpoint and get the list of students.

fix on the composer file so it include the .env in the container volume

Component to list all students added and a dialog to show their individual attendances (pending use the api)

default layout added to the App.vue

route to access StudentView added to the index.js router file